Changed key release countdown from over(...) to remember(Until.Expired);
Although it would be safer to apply the countdown on setting the
duration, and release keys when the item expires, but
self-bondage test-sessions revealed that keys are released when the item
expires, and not when it's removed.

Requesting the countdown manually also results in an automatic release
(when the script doesn't progress anymore and the hold duration timer
reaches zero), so the key release device is still safe to use.